struct e empilhar

+++ MATD04 - ESTRUTURAS DE DADOS 2022.1 FÉRIAS/EQUIPE 4/Trabalho Final Stack/functions.h
@@ -8,7 +8,7 @@ typedef struct NO{
 
 typedef struct PILHA{
   No *topo;
-  float tamanho;
+  int tamanho;
 } Pilha;
 
 void inicializarPilha(Pilha *pilha) {
